fix: set port to `u16`

+++ src/config.rs
@@ -1,7 +1,7 @@
 #[derive(Debug, Clone)]
 pub struct Config {
     pub server_address: String,
-    pub server_port: i32,
+    pub server_port: u16,
     pub database_url: String,
     pub jwt_secret: String,
     pub jwt_expires_in: String,
